fastapi_tasks/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
asyncmy==0.2.9
asyncpg==0.30.0
bcrypt==3.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
coverage==7.6.4
cryptography==43.0.3
databases==0.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Mako==1.3.6
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tomli==2.0.2
typing_extensions==4.12.2
uvicorn==0.32.0